Renamed `BenchmarkKey` into `BenchmarkKeyUtils`, removed unused methods.
