+++ platform-sdk/swirlds-benchmarks/src/jmh/java/com/swirlds/benchmark/BaseBench.java
@@ -119,7 +119,7 @@ public abstract class BaseBench {
 
         verify = benchmarkConfig.verifyResult();
 
-        BenchmarkKey.setKeySize(keySize);
+        BenchmarkKeyUtils.setKeySize(keySize);
 
         // recordSize = keySize + valueSize
         BenchmarkValue.setValueSize(Math.max(recordSize - keySize, RECORD_SIZE_MIN));
